DEP: cluster: deprecate `vq.py_vq` and tidy public API
[skip ci]
