+++ scipy/cluster/hierarchy/_hierarchy.pyx
