+++ scipy/cluster/hierarchy/_hierarchy_distance_update.pxi
