+++ scipy/cluster/hierarchy/_optimal_leaf_ordering.pyx
