+++ scipy/cluster/hierarchy/_structures.pxi
