+++ scipy/cluster/hierarchy/tests/hierarchy_test_data.py
