+++ scipy/cluster/hierarchy/tests/test_disjoint_set.py
