+++ scipy/cluster/hierarchy/tests/test_hierarchy.py
@@ -43,9 +43,11 @@ from scipy.cluster.hierarchy import (
     correspond, is_monotonic, maxdists, maxinconsts, maxRstat,
     is_valid_linkage, is_valid_im, to_tree, leaves_list, dendrogram,
     set_link_color_palette, cut_tree, optimal_leaf_ordering,
-    _order_cluster_tree, _hierarchy, _EUCLIDEAN_METHODS, _LINKAGE_METHODS,
     complete, average, weighted, centroid, median)
-from scipy.cluster._hierarchy import Heap
+from scipy.cluster.hierarchy._hierarchy import Heap
+from scipy.cluster.hierarchy._hierarchy_impl import (
+    _order_cluster_tree, _hierarchy, _EUCLIDEAN_METHODS, _LINKAGE_METHODS
+)
 from scipy.spatial.distance import pdist
 from scipy._lib._array_api import (eager_warns, make_xp_test_case,
                                    xp_assert_close, xp_assert_equal,
