+++ scipy/cluster/vq/_vq.pyx
