+++ scipy/cluster/vq/_vq_impl.py
@@ -1,69 +1,3 @@
-"""
-K-means clustering and vector quantization (:mod:`scipy.cluster.vq`)
-====================================================================
-
-Provides routines for k-means clustering, generating code books
-from k-means models and quantizing vectors by comparing them with
-centroids in a code book.
-
-.. autosummary::
-   :toctree: generated/
-
-   whiten -- Normalize a group of observations so each feature has unit variance
-   vq -- Calculate code book membership of a set of observation vectors
-   kmeans -- Perform k-means on a set of observation vectors forming k clusters
-   kmeans2 -- A different implementation of k-means with more methods
-           -- for initializing centroids
-
-Background information
-----------------------
-The k-means algorithm takes as input the number of clusters to
-generate, k, and a set of observation vectors to cluster. It
-returns a set of centroids, one for each of the k clusters. An
-observation vector is classified with the cluster number or
-centroid index of the centroid closest to it.
-
-A vector v belongs to cluster i if it is closer to centroid i than
-any other centroid. If v belongs to i, we say centroid i is the
-dominating centroid of v. The k-means algorithm tries to
-minimize distortion, which is defined as the sum of the squared distances
-between each observation vector and its dominating centroid.
-The minimization is achieved by iteratively reclassifying
-the observations into clusters and recalculating the centroids until
-a configuration is reached in which the centroids are stable. One can
-also define a maximum number of iterations.
-
-Since vector quantization is a natural application for k-means,
-information theory terminology is often used. The centroid index
-or cluster index is also referred to as a "code" and the table
-mapping codes to centroids and, vice versa, is often referred to as a
-"code book". The result of k-means, a set of centroids, can be
-used to quantize vectors. Quantization aims to find an encoding of
-vectors that reduces the expected distortion.
-
-All routines expect obs to be an M by N array, where the rows are
-the observation vectors. The codebook is a k by N array, where the
-ith row is the centroid of code word i. The observation vectors
-and centroids have the same feature dimension.
-
-As an example, suppose we wish to compress a 24-bit color image
-(each pixel is represented by one byte for red, one for blue, and
-one for green) before sending it over the web. By using a smaller
-8-bit encoding, we can reduce the amount of data by two
-thirds. Ideally, the colors for each of the 256 possible 8-bit
-encoding values should be chosen to minimize distortion of the
-color. Running k-means with k=256 generates a code book of 256
-codes, which fills up all possible 8-bit sequences. Instead of
-sending a 3-byte value for each pixel, the 8-bit centroid index
-(or code word) of the dominating centroid is transmitted. The code
-book is also sent over the wire so each 8-bit code can be
-translated back to a 24-bit pixel value representation. If the
-image of interest was of an ocean, we would expect many 24-bit
-blues to be represented by 8-bit codes. If it was an image of a
-human face, more flesh-tone colors would be represented in the
-code book.
-
-"""
 import warnings
 import numpy as np
 from collections import deque
@@ -71,17 +5,19 @@ from scipy._lib._array_api import (_asarray, array_namespace, is_lazy_array,
                                    xp_capabilities, xp_copy, xp_size)
 from scipy._lib._util import (check_random_state, rng_integers,
                               _transition_to_rng)
+from scipy._lib.deprecation import _deprecated
 from scipy._lib import array_api_extra as xpx
 from scipy.spatial.distance import cdist
 
 from . import _vq
 
-__docformat__ = 'restructuredtext'
-
-__all__ = ['whiten', 'vq', 'kmeans', 'kmeans2']
+__all__ = ['ClusterError', 'kmeans', 'kmeans2', 'py_vq', 'vq', 'whiten']
 
 
 class ClusterError(Exception):
+    """
+    An ``Exception`` raised during clustering.
+    """
     pass
 
 
@@ -221,7 +157,7 @@ def vq(obs, code_book, check_finite=True):
     return py_vq(obs, code_book, check_finite=False)
 
 
-def py_vq(obs, code_book, check_finite=True):
+def _py_vq(obs, code_book, check_finite=True):
     """ Python version of vq algorithm.
 
     The algorithm computes the Euclidean distance between each
@@ -274,6 +210,13 @@ def py_vq(obs, code_book, check_finite=True):
     code = xp.argmin(dist, axis=1)
     min_dist = xp.min(dist, axis=1)
     return code, min_dist
+
+
+_py_vq_dep_msg = (
+    "`scipy.cluster.vq.py_vq` was unintentionally public,"
+    " use `scipy.cluster.vq.vq` instead."
+)
+py_vq = _deprecated(_py_vq_dep_msg)(_py_vq)
 
 
 def _kmeans(obs, guess, thresh=1e-5, xp=None):
