+++ scipy/cluster/vq/tests/meson.build
@@ -1,0 +1,10 @@
+python_sources = [
+  '__init__.py',
+  'test_vq.py'
+]
+
+py3.install_sources(
+  python_sources,
+  subdir: 'scipy/cluster/vq/tests',
+  install_tag: 'tests'
+)
